Support removing cinder-backup from pcmk control
TripleO traditionally deploys the cinder-backup service to run
A/P (active/passive) under pacemaker, but cinder itself supports
running A/A. This patch adds deploy_steps_tasks to remove the
service from pcmk control (essentially deleting the pcs resource)
prior to deploying it as a regular containerized (non-pcmk) service.

A new cinder-backup-active-active.yaml environment file helps the
user clarify their intent. The legacy cinder-backup.yaml file
continues to deploy the service under pacemaker (to preserve the
default deployment behavior), but rather than encouraging users to
edit the file to switch deployment modes (as a comment in the file
used to advise), users can proactively deploy the A/A mode (or switch
from A/P to A/A) by including the new environment file.

Change-Id: Ic94025f51083da77f4ba6ee934ebcf32bcb9909f