Deploy RGW by default when cephadm environment is included
This patch just includes the content of ceph-rgw.yaml in the common
cephadm environment file. This means that when Ceph is deployed by
TripleO, the RGW component will be applied using cephadm.
The cephadm-rbd-only environment file is also added and can be applied
for all the use cases where swift is used and rgw shouldn't replace
it.

Change-Id: I5e841ea76575eed28f1c8fc9a95cd4d3fd62dbc4