Disable global_id reclaim for Ceph mons in scenario001
As per [1], patched monitors raise the AUTH_INSECURE_GLOBAL_ID_RECLAIM_ALLOWED
health alert if auth_allow_insecure_global_id_reclaim is enabled.
This is ok for general deployments, but it's not required in CI deployments.
Adding this option to the mon section using the CephConfigOverrides
approach ensures the cluster stay in health_ok.

[1] https://docs.ceph.com/en/latest/security/CVE-2021-20288/

Change-Id: Ib8d3f62d233ce5a42e9923547c08add6e5859cfb