Add OVNEncapType option to the ovn controller template
Since OVN 2.13 it is possible to use vxlan encapsulation in the OVN
tunnel networks.
To do that, it has to be set in the external_ids:ovn-encap-type. There
is support for that in the puppet-ovn module already.
This patch adds missing TripleO-heat-template variable.

Related-bz: #1881697
Change-Id: I54e617745e76f27a4c88156a9b1b17d0255ec837
(cherry picked from commit 1542df355bd22b83baaba1444fc58028046a5185)