Sync full /etc/leapp/files directory.
To make the leapp upgrade procedure a little bit simpler the two files
required by leapp to work, repomap.csv and pes-events.json, were being
retrieved from the Undercloud node to the Overcloud nodes. However, the
leapp tool now requires two new files to work properly, which are not
being considered in the upgrade tasks, making the upgrade for the
overcloud nodes fail.

To avoid that the addition of new file requirements will break the
Overcloud leapp upgrade, this patch synchronizes the whole content
under /etc/leapp/files from the Undercloud to the Overcloud nodes using
find, fetch and copy.

Change-Id: I765dc048016bb92462ccd55132e48a61ef8a9ecd
Resolves: rhbz#1953540