Default CephClientConfigVars within --working-dir
As per I0d803f695c725c58ef2e6b655753b6c8248d1b2f, all files
created by the overcloud deploy command should be in the
tripleo working directory instead of using tmp dirs or files
directly in $HOME.

Change the default of CephClientConfigVars so that it doesn't
genereate the ceph_client.yml in /home/stack but instead puts
it in the cephadm directory within config-download as per the
value of {{ playbook_dir }} when ansible evaluates it. This
will result in the ceph_client.yml also being within the working
directory.

Change-Id: Ia6829fb8a85776878b737d1ae2c5567c2fd2ed1e