Add option for enabling rsyslog reopenOnTruncate
reopenOnTruncate ensures rsyslog immediately recognizes when a logrotation happens on a file. When not on, rsyslog will wait for a log file to fill to its original capacity before consuming any additional logs.

Setting this here will ensure that every service configured to work with rsyslog will have this parameter. Default to off because this is an experimental feature

Depends-On: https://review.opendev.org/c/openstack/puppet-tripleo/+/790154
Change-Id: Iba6a0da80cb75a93537a014dba8e77032101e570