[ffwd][train-only] Rebuild clouds.yaml before running keystone endpoint configuration.
In [0] the rebuilding of Keystone endpoint configuration was added
when running the FFWD workflow. However, it was not considered that
the Undercloud upgrade only adds the content for the undercloud stack
in clouds.yaml, being the Overcloud's stack information missing from
the file. This provokes the task to fail while trying to create
the Keystone endpoints, as no cloud matching the cloud's stack name
is found.

This patch anchors the block in charge of generating the clouds.yaml
content during the external_deploy_tasks and calls it in the step 0
so that we ensure the file is ready when executing the post_upgrade
task in step 1.

[0] - Ic1e63ca773fb20f10f16fc3835f8f263cb3b6929

Change-Id: Ie014961438f7c255a4eb9907596e873645bfcd46
Resolves: rhbz#1960579