Move ceph_admin_extra_vars logic to tripleo-ansible
Instead of creating ceph_admin_extra_vars with
distribute_private_key always set to true, set
that variable to true only when appropriate based
on logic in the depends-on patch.

Also, it is not necessary to override the values of
tripleo_admin_generate_key or ssh_servers to create
the ceph-admin user for cephadm.

Related-Bug: #1928717
Depends-On: I8343c419c140670f01bdc94b4c8130004bac64e1
Change-Id: I2bacf82f85e5c78f5ae603460919cf3ff7130e9c