[Train-Only] Add leapp answer authselect_check.confirm.
There is a new tool called authselect in RHEL8 that replaced authconfig.
When upgrading TLS-Everywhere environments the PAM files seem to be modified
out of authconfig, which triggers the authselect_check actor to ask the
operator if its desired to migrate from authconfig to authselect.
The migration to authselect should be safe, for that reason we are assigning
the constant value of True in the answers file before performing the Leapp
upgrade.

Change-Id: I8b7a1ee8be6ce0e72b08db5f354a458bdcf4ef57
Resolves: rhbz#1952574