Add environment for external network VIP
This environment allows us to control the external VIP for Undercloud or Standalone SSL.
When used, a network data has to be provided (e.g. network_data_undercloud.yaml) and
`openstack tripleo deploy` must be called with --public-virtual-ip.

It's not really needed for Undercloud because
environments/undercloud.yaml already sets the resource right, but for
Standalone users this file will be useful when they want public
endpoints to be listening on another network.

Change-Id: I461e0fdc8fc717beed3aea7a66e41de1e11116da