[train-only] Set leapp parameters to target RHEL release
When enacting the upgrade from RHEL7 to RHEL8 using leapp,
the default behaviour for leapp is to upgrade to the latest
RHEL version it supports. This is currently 8.4, but will
end up being RHEL 8.6 in the future.

To ensure we get the correct target RHEL release, we set the
UpgradeLeappDevelSkip value during upgrade-prepare. We also
ensure that these valuiies are not set for ceph nodes, because
they are not release locked like OSP is.

This is train-only because it is unique to the RHEL7 to
RHEL8 upgrade process.

Related: rhbz#1962365

Change-Id: I33daab14f877ca063021ee53424ae180a7ef0966