Remove ENABLE_* environment about puglins not installed
Currently we maintain list of installed plugins and not installed
plugins in horizon template but the list has some incorrect/invalid
items. Futhermore kolla sets ENABLE_* to no by default unless
the environments are explicitly set so there is no need to maintain
"uninstalled" items.

This change simplifies the list of plugins to include only installed
and enabled plugins.

Change-Id: I3e7532e0e33326f1e4ff8f59bafe41d8582abd03