Make UpgradeLeappDevelSkip per-role
When specific roles are deployed, for example CephStorage,
then the parameters for UpgradeLeappDevelSkip are different
to all other roles.

This change makes UpgradeLeappDevelSkip a per-role parameter.

Included in this is the appropriate tagging for the other
parameters which had the same work done in:

- https://review.opendev.org/779522
- https://review.opendev.org/774537

Release notes are included for all of the above too.

Related-Bug: rhbz#1962365

Change-Id: I522951cc7ec6034bb6287dd144a9fc433c0090e0