[train-only] Remove unsupported kernel modules before leapp
When enacting the upgrade from RHEL7 to RHEL8 using leapp,
it will hard-stop if these modules are loaded. Thus far,
these are the ones tripping up CI and some customers.

To cater to customers who might hit this, and to resolve
issues in CI, we include a new parameter
LeappUnloadKernelDrivers which contains a list of the Kernel
drivers we are required to unload before running leapp upgrade.

This parameter defaults to ['floppy', 'pata_acpi'] which are
the ones impacting our CI jobs so far, but the existence of
this parameter in combination with the RoleParametersValue
would allow us to set a different type of drivers per role,
if needed.

For all the drivers appearing in that list, an unloading
will be executed via modprobe Ansible module right before
executing the leapp upgrade command per overcloud node.

Related: rhbz#1962365

Change-Id: I3d32857e783a9ef6685896ae9f77494b4c6ed2df