Get roles data from stack output on UC upgrade
Since not all information can be exported from the
stacks, the intent is to backport the depends-on
to the release intended as the upgrade from release.
i.e the depends on need to go back to Train.

The alternative is to require the user to provide the
roles_data file for each overcloud stack managed by
the undercloud being upgraded as input to the upgrade
command. Support for this is provided in the previous
patch in this chain.

Depends-On: https://review.opendev.org/791476
Related: blueprint network-data-v2-ports
Change-Id: Icc6a7a438e9d0f39d003d1cf8ed84d6fb1d5485a
(cherry picked from commit ff48f2ac1eef50b575570a90ae417145dd1e9935)