Limit collectds memory usage
to guard nodes from running out of memory.

If there is a connection issue with collectd not being able to
write to event or metrics endpoints, this can lead to accumulating
to large portions of memory. Limiting the memory usage via podman
memory will prevent the container to grow limitless. In case of
using more than 512 MB, podman will restart the container.

Change-Id: I0b65b2c7878fe5ebd0d7f1a6f2ba8df23f2f0b28