Set tripleo_cephadm rbd mirror vars
This patch introduces the logic of adding and setting the rbd mirror
parameters required, later in the process, by cephadm.
By doing this we can not only deploy the rbd-mirror daemon (via the
cephadm spec file), but also run a set of ceph commands to properly
configure mirroring on a given pool.

Depends-On: I0d1ce33270ccf8636925c3534812c3b1cf650103
Change-Id: I7b20399d55c3de82c8aeddc3317c9972409217da