Distribute iscsid.conf to all containers using iscsi
This patch updates the way files related to iscsi are distributed
to the cinder, glance and nova containers that use the protocol.

Previously it was thought that only the iscsid container needs
access to /etc/iscsi/iscsid.conf, but the LP bug reveals the client
side also reads the file in order to determine the list of chap
algorithms to offer when initiating an iscsi connection.

The bug was exposed when testing a secure environment that uses a
non-default list of chap algorithms. The iscsid container was using
the customized list, but the client containers (e.g. nova) were
using the default list, which caused iscsid to reject connections.

Closes-Bug: #1932181
Change-Id: Iad255451726867dc172404513fdac4ad0599c4c0
(cherry picked from commit 48fd886a03b8d8249b6ea539b76bbd50e081a85e)