Always run network config for undercloud
We need to ensure that os-net-config is run on upgrades. On the
undercloud we want the os-net-config to always run so let's set the
NetworkDeploymentAction correctly.

NOTE: Changed Change-Id as we incorrectly early cherry-picked and
and to revert.
Original-Change-Id: I749714c1f0af3787020f3e9b88f752d272b91d37
Change-Id: Id3924a027c9bd7865c9fb081a2eda4f8be069a57
Related-Bug: #1930151
(cherry picked from commit 3b043e4953393c7723553c84cc094a25902cd083)