Introducing the ComputeVdpa role
Now that nova supports libvirt's vDPA implementation [1][2], we need
deployment code to enable such a feature.

The ComputeVdpa role is basically a copy of ComputeSriov with the
addition of the ExtraKernelModules. Beside these modules, operators
should define to their ``NovaPCIPassthrough`` as they normally do
for standard SR-IOV as a role parameter.

When having the right modules loaded, at least in RHEL8.5, the kernel
will automatically create /dev/vhost-vdpa-X devices for each one of the
VFs.

[1] https://opendev.org/openstack/nova-specs/src/branch/master/specs/wallaby/implemented/libvirt-vdpa-support.rst
[2] https://bugzilla.redhat.com/1873409

Depends-On: https://review.opendev.org/c/openstack/os-net-config/+/798926
Related: https://bugzilla.redhat.com/1956297
Change-Id: Ifda3ac08c596f65bc75d869097a0f72e816ba783