Add DeployedCeph boolean parameter
If the Ceph cluster has already been deployed but needs to be
configured so that the overcloud can use it (create cephx keys,
pools, configure RGW with haproxy, etc), then this parameter
should be set to true. Set this parameter to false to have
cephadm deploy the ceph cluster during overcloud deployment.
DeployedCeph and CephDynamicSpec are mutually exclusive.

Depends-On: Id32f1135c857eb4ccc4a6c7356957ff0f5182839
Change-Id: Ic80f5d6fd0956040b9ced1e9e0f6079a2c0d3049