vncproxy: removal of tags and steps change from post_upgrade_tasks
Post upgrade tasks are limited to 4-1 steps, so these
tasks are not executed.

Deleting nova-consoleauth is failing with some of the CIs and this is
mostly a best-effort kind of task, so adding ignore_errors.

Also, replacing some yes/nos with boolean as it's best practice.

Related: https://bugzilla.redhat.com/1921112
Change-Id: I0bc39170569c42a0be31233fd7ec068613547790