Remove NovaVncProxyNetwork from ServiceNetMap
In Icef2481b65b41b524ad44eeecfbee4451006e1d2 we moved to use
NovaLibvirtNetwork as the single network to configure instance console
components.
Due to how hieradata is being rendered via tripleo-hieradata, the
removed NovaVncProxyNetwork resulted in no-vnc-proxy service to
fall back to the control plane network because NovaVncProxyNetwork
was removed.
NovaVncProxyNetwork was then reintroduced with
I595294956a4a7a0e2280c685ac1d2543edbc32f2 but as a side effect
again introduced the inconsistency on the console configuration.

This patch removes the NovaVncProxyNetwork from ServiceNetMap
but set it as part of service_net_map.j2.yaml and keep it in
sync with NovaLibvirtNetwork.

Change-Id: I08fa1ee947574e6fac24885b17706a22054fcba9
Closes-Bug: #1917719