Neutron: Use the pids subsystem to look up container id
Currently we rely on cgroup name to detect id of sidecar containers but
the name is not managed by podman and it is not guaranteed that it is
consistent with container id.
This changes switches the base record to the pid subsystem which is
actively managed by podman and more likely to be consistent with
container ids.

Closes-Bug: #1933918
Change-Id: Id57f86e62409e257f5619dc8b360cee3f922c805