Do not fail if haproxy container UUID changes
There is a possibility that haproxy container crashed and restarted
with another UUID, while the deployment is trying to reload it
after reconfiguring the certificates, by sending a
`kill -HUP <uuid-of-the-container>`. In that case, ingore errors
for the commands block, since there is no longer need to reload the
newly created container.

Related rhbz#1973674

Change-Id: I4b40d73ab329dc219ee7a387201b0747a6233ed4
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>