Add heat-api environment
Add an environment file that can be used to enable Heat API. While the
service is already enabled by default, there are other environments that
could be used to disable it (such as standalone-tripleo.yaml). Having
this environment makes the service easy to re-enable without having to
use a custom environment file.

Change-Id: Ib60631a1c17786fa6d4aa36da03b819846144806
Signed-off-by: James Slagle <jslagle@redhat.com>