Drop CollectdAmqpInstances from collectd-write-qdr.yaml
Drop the CollectdAmqpInstances parameter from collectd-write-qdr.yaml in
order to allow an administrator to override the default specified
topics. If the administrator of the cloud used collectd-write-qdr.yaml
and then needed to override the topics, a merge of the defaults and
custom topics would be added to the 10-amqp.conf for collectd, resulting
in data being published to unused topics.

Change-Id: I554d8677156934c4d49acb1346afea16339ad5c2
Signed-off-by: Leif Madsen <lmadsen@redhat.com>