Ensure startup files exist before containers might be started
Change Ib3476e53f89b50bae72b9c95a5d3dec51ed3de7e moved Octavia external
deployment tasks to post deployment which occurs after the containers
are launched. This causes them to fail on startup. While deployment
continues, configuration is generated and the containers started later,
it gives the appearance that the containers are failing at some point
during deployment.

This patch creates and empty file so the file passed in the command line
exists when the container is first started. The caveat is that the
configuration is incomplete until after the post deploy tasks have run.

Related-bug: #1863595
Change-Id: Id26edd4e7414c35cc0d2b4d743978a0d56c58fee