Remove libvirt package from overcloud nodes
Now that nova_libvirt is containerized, there is no need
to have libvirt installed in the overcloud nodes. In fact,
if it is installed on the overcloud nodes it causes issues
the virtual machines on the host to shut down when doing
an update on the train releases due to a miscomunication
between the container's services and the host ones.

This patch uses the to_remove mechanism from leapp to remove
the libvirt package during the leapp upgrade. This ensures
that leapp removes the package during its process. This is
the least risky time to remove the package and should help
save some time during the leapp process.

Related: rhbz#1963164

Change-Id: I8ab42f5dde132875c169db6e69b7d0db2a2b6cc4