Parameter to set api policy to limit live migration to role
Adds NovaRestrictLiveMigration boolean parmeter to enable an
api policy to allow live migration only for a specific role
which can be customized using NovaRestrictLiveMigrationRole.
This feature is to prevent the default admin role users to
be able to use live migration on coincidence.
Additional policies specified using NovaApiPolicies get
merged with this polic

Depends-On: https://review.opendev.org/c/openstack/puppet-nova/+/802251
Change-Id: If0b60037f7bb7084d0d5e549670e9d8bf53330a8