Set stop grace periods for designate services
Many of the designate services perform background syncrhonization
tasks and can lag a bit on shutdown, especially on heavily loaded or
large scale systems. Increasing the stop grace periods reduces the
likelihood that the system might get into a non-consistent state or a
state where a great deal of reconciling needs to occur on restart.

Change-Id: If521662eb4c47ada45cf0a9edd0a58db8a1d038d