[train-only] Add the ability to remove leapp actors
It may be necessary to remove specific actors from the leapp
process to unblock the upgrades. An example included in this
patch is the persistentnetnamesdisable which needs to be
removed in order to prevent requiring a reboot after renaming
the network interfaces.

In this patch we add the ability to define actors to be removed
just before running leapp. The hook is role-specific to allow
maximum flexibility.

Resolves: rhbz#1984875

Change-Id: Id9c141b19383505c1a95948823722822b814bde1