Swap xinetd/tftp-server for dnsmasq to support ironic-pxe operations
tftp-server and xinetd are not anticipated to be in RHEL9 and as
such we need to remove their use in order to facilitate a transition
to RHEL9 when the time comes.

In this case, we're anticipating we can just directly launch dnsmasq
instead of in.tftpd directly within the container.

Change-Id: Ie75aa0f2994a4cf662905b912231a12e3bd8afe7