Eliminate unnecessary role looping in deploy-steps.j2
These role task inclusions force a linear-like sequence
and result in a lot of unnecessary skipped tasks. This
takes extra time, and causes a lot of log noise which
makes debugging more difficult.

Given that a host can ever only have a single role, it
makes more sense to just include the applicable role
task file if it exists.

Change-Id: Id2cadceaaf563dc94fcf2d8ce0f0cb9054a65d40