Consider GlanceMultistoreConfig when setting glance_pool
With glance supporting multiple stores we need to update glance_pool
so that it explores the content of GlanceMultistoreConfig to see if
any of the configured backends are of rbd type.

We also add a the CephClusterName parameter into the CI rbd_store
multistore for Glance, which is a mandatory parameter to correctly
set the permissions on the keyring files of every Ceph cluster.

Co-Authored-By: Yatin Karel <ykarel@redhat.com>
Change-Id: Iee670befb2bcbaa2b666c6b99f5153391b62049e
(cherry picked from commit 151ca15e81799606972d9fe7ad65107dfde5a427)