Adjust HAProxy script to allow for ID change
It is possible that the UID of the container
changes between our first parse and the script
we execute to copy the TLS Cert. We can re-check
the container ID in the script to be safe.

Also, we need to fail appropriately if we can't
find the container, or can't restart it.

Related rhbz#1973674
Closes-Bug: #1940729
Change-Id: I1b8c8e83d7b4a14a8643d63a61519f6bbac5b3d6

Change-Id: Ifafb2e71da1a921eeba8d8c6197cfb74d1ee045e