Disable cephadm when ceph is deployed
After ceph is deployed, which means day1 operations are over
and the Ceph cluster daemons are up && running, cephadm can
be paused by running the 'disable_cephadm' playbook.
This can be triggered at step3 (to ensure all the daemons are
started) and can be triggered by the DisableCephadm exposed
parameter.

Closes-Bug: 1940866
Depends-On: I5952bb8a7a327e37a39acc95751cd5746fd9a41d
Change-Id: Idbafdaced85c945d38940be83c1a269c73e4cb6b