Add a global configuration option for secure RBAC
This commit adds a single option that we can use to turn secure RBAC on
in all the OpenStack services. This is a contrasting approach to using
service-specific variables.

We don't intend anyone to use this functionality until we can update
each OpenStack service to use the correct token when communicating with
other services. Deployers should also only use this if they are sure all
services in their deployment actually support secure RBAC, which might
not be for a while, but this allows us to lay down the plumbing in each
service to turn it on now, which will be useful for testing and flushing
out what we need to change in TripleO as well as the upstream services.

Change-Id: I13262a4f1a6e850d66b2c687e730e0c2004c1f29
(cherry picked from commit 57a8efd4077d4ba7c39a48834eac9e14e78edc08)