Fix the CephDashboardNetwork default
The Ceph Dashboard uses by default the Grafana network to bind the
overcloud backends to the specified port.
However, the CephDashboardNetwork parameter is used to define the
haproxy frontend and the related vip, used to reach the Ceph backends.
The purpose of this patch is to make sure the right defaults are
used for TripleO deployed Ceph clusters.

Depends-On: I716f94b4e9eae4f26b9ab11c4f412aebcb0b58c2
Change-Id: I1dd5581317462f638185c918c85c86f953ddf28f