[Wallaby only] Drop unused openstack DB users in mysql
Openstacklib was used in a way that generated additional
users in mysql which are never used. In order to fix
this user creation on an existing environment, we must
delete the unused users manually, as puppet-mysql never
deletes users in the DB.

Add an upgrade task in Wallaby only because we stopped
generating unused users in Master. Also adjust the
indentation to avoid whitespace diff to would complicate
backports from Master to Wallaby.

Change-Id: Ifc2d65b1daa245af3fa97154822cf9f8412d259d