Wallaby and older: Remove heat-dashboard parameters from local_settings
This change ensures the parameters for heat-dashboard are removed
from the base local_settings, because the parameters are now configured
in the independent setting files.

Note this change is required only in Wallaby and older and since Xena
the parameters are no longer included in the base local_settings.

Conflicts:
	deployment/horizon/horizon-container-puppet.yaml

Depends-on: https://review.opendev.org/815362
Depends-on: https://review.opendev.org/816330
Change-Id: I1b1ffdd68501d61a03d8153c4fc538b6d26c8344
(cherry picked from commit ff66c15547e6b1dbfe31d79b5852099ba16384ac)